moved data out of src same with the db
