+++ app/src/services/database/db_friends.ts
@@ -17,6 +17,7 @@ export async function createFriendRequest(
 	if (result.lastID === undefined) {
 		throw new Error('Failed to retrieve the last inserted ID.');
 	}
+	db.close();
 	return result.lastID;
 }
 
@@ -29,6 +30,7 @@ export async function acceptFriendRequest(requestId: number): Promise<void> {
 		'UPDATE friends SET accepted = true WHERE id = ?',
 		requestId
 	);
+	db.close();
 	if (result.changes === 0) {
 		throw new Error('No friend request found to accept.');
 	}
@@ -40,6 +42,7 @@ export async function rejectFriendRequest(requestId: number): Promise<void> {
 		driver: sqlite3.Database,
 	});
 	const result = await db.run('DELETE FROM friends WHERE id = ?', requestId);
+	db.close();
 	if (result.changes === 0) {
 		throw new Error('No friend request found to reject.');
 	}
@@ -57,6 +60,7 @@ export async function removeFriendship(
 		'DELETE FROM friends WHERE ((requester_id = ? AND requested_id = ?) OR (requester_id = ? AND requested_id = ?))',
 		[userId, friendId, friendId, userId]
 	);
+	db.close();
 }
 
 export async function getFriends(userId: number) {
@@ -64,7 +68,7 @@ export async function getFriends(userId: number) {
 		filename: dataBaseLocation,
 		driver: sqlite3.Database,
 	});
-	return await db.all(
+	const friends = await db.all(
 		`SELECT u.id, u.username, u.displayname
 		FROM users u
 		INNER JOIN friends f
@@ -73,6 +77,8 @@ export async function getFriends(userId: number) {
 		WHERE f.accepted = true`,
 		[userId, userId]
 	);
+	db.close();
+	return friends;
 }
 
 export async function getFriendRequest(
@@ -83,10 +89,12 @@ export async function getFriendRequest(
 		filename: dataBaseLocation,
 		driver: sqlite3.Database,
 	});
-	return await db.get<Friend>(
+	const friend = await db.get<Friend>(
 		`SELECT * FROM friends WHERE requester_id = ? AND requested_id = ? OR requester_id = ? AND requested_id = ?`,
 		[fromUserId, toUserId, toUserId, fromUserId]
 	);
+	db.close();
+	return friend;
 }
 export async function getFriendRequestById(
 	requestId: number
